+++ blender3d_intergration/blender_python/blender_commands_generator.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-from blender3d_intergration.blender_python_working.blender_python_commands import BlenderPythonCommands as bpy
+from blender3d_intergration.blender_python.blender_python_commands import BlenderPythonCommands as bpy
 
 from blender3d_intergration.enums import FileExtensions
 from blender3d_intergration.trajectories_api.models import Trajectory, CoordinatesInTime
